refactored CSS file structure for better coding practices
