+++ public/log.txt
@@ -564,3 +564,5 @@ March 9
 Finalized SQL integration for user authentication and data storage.
 Ensured seamless linkage between Appwrite and SQL database.
 Conducted final test cases for authentication and storage.
+March 11
+
